authentication

+++ backend/controllers/question_user_answer.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const question_user_answer = require('../src/models/question_user_answer');
 const QuestionUserAnswerModel = require('../src/models/question_user_answer');
+var jwt = require('jsonwebtoken');
 //var db = require('../index')
 
 
@@ -22,6 +23,24 @@ router.use(express.urlencoded({
 router.post('/add', function (req, res) {
     let request = req.body
     console.log('request received');
+
+    let token = req.headers.authorization
+    console.log('token : ' + token);
+    if (!req.headers.authorization) {
+        return res.status(403).json({ error: 'No credentials sent!' });
+    }
+    try {
+        var decoded = jwt.verify(token, 'thebestsecretkeyever');
+        console.log(decoded['data'])
+        let jwtData = decoded['data'];
+        if(jwtData['role']!=='Admin'){
+            return res.status(403).json({ error: 'Insufficient access.' });
+        }
+    } catch (err) {
+        console.log(err) // bar
+        return res.status(403).json({ ...err });
+    }
+
     QuestionUserAnswerModel.create({
         userId: request['userId'],
         answerId: request['answerId'],
@@ -44,6 +63,25 @@ router.post('/add', function (req, res) {
 
 router.post('/delete', async function (req, res) {
     let request = req.body
+
+
+    let token = req.headers.authorization
+    console.log('token : ' + token);
+    if (!req.headers.authorization) {
+        return res.status(403).json({ error: 'No credentials sent!' });
+    }
+    try {
+        var decoded = jwt.verify(token, 'thebestsecretkeyever');
+        console.log(decoded['data'])
+        let jwtData = decoded['data'];
+        if(jwtData['role']!=='Admin'){
+            return res.status(403).json({ error: 'Insufficient access.' });
+        }
+    } catch (err) {
+        console.log(err) // bar
+        return res.status(403).json({ ...err });
+    }
+
     await QuestionUserAnswerModel.destroy({
         where: {
             userId: request['userId'],
@@ -58,7 +96,25 @@ router.post('/delete', async function (req, res) {
 
 
 router.get('/countAll', async function (req, res) {
-    QuestionUserAnswerModel.count({})
+
+    let token = req.headers.authorization
+    console.log('token : ' + token);
+    if (!req.headers.authorization) {
+        return res.status(403).json({ error: 'No credentials sent!' });
+    }
+    try {
+        var decoded = jwt.verify(token, 'thebestsecretkeyever');
+        console.log(decoded['data'])
+        let jwtData = decoded['data'];
+        if(jwtData['role']!=='Admin'){
+            return res.status(403).json({ error: 'Insufficient access.' });
+        }
+    } catch (err) {
+        console.log(err) // bar
+        return res.status(403).json({ ...err });
+    }
+
+        QuestionUserAnswerModel.count({})
         .then(function (count) {
             console.log(count)
             res.json({
@@ -70,6 +126,24 @@ router.get('/countAll', async function (req, res) {
 });
 
 router.get('/countByUser/:id', async function (req, res) {
+
+    let token = req.headers.authorization
+    console.log('token : ' + token);
+    if (!req.headers.authorization) {
+        return res.status(403).json({ error: 'No credentials sent!' });
+    }
+    try {
+        var decoded = jwt.verify(token, 'thebestsecretkeyever');
+        console.log(decoded['data'])
+        let jwtData = decoded['data'];
+        if(jwtData['role']!=='Admin'){
+            return res.status(403).json({ error: 'Insufficient access.' });
+        }
+    } catch (err) {
+        console.log(err) // bar
+        return res.status(403).json({ ...err });
+    }
+
     const id = req.params.id;
     QuestionUserAnswerModel.count({
         where: {
